requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
python-dotenv
alembic
python-jose[cryptography]
passlib[bcrypt]
pytest
httpx
